pyproject.toml
[project]
name = "pydantic-deepagents-chutesai"
version = "0.1.0"
description = "Pydantic Deep Agents configured for chutes.ai provider"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Standard Syntax"},
]
dependencies = [
    "pydantic-deep>=0.2.9",
    "pydantic-ai>=0.1.0",
    "pydantic>=2.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.8.0",
    "pyright>=1.1.0",
]
sandbox = [
    "pydantic-ai-backend[docker]>=0.0.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "Q"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
